Change the "registrationSource" of the embedded signup form
This is just one little tweak that I forgot about when I was giving you the snippet of code earlier. This changes the `registrationSource` variable in your signup form embed code to `gsa-auctions`. The purpose of this is so we can better identify where our user registrations are coming from. The value can really be anything, we just want an easy identifier to uniquely identify the different web pages where our signup form is being embedded.